refactor/package/model repo add
