+++ src/main/java/com/ghbt/ghbt_starbucks/api/product/service/ProductServiceImpl.java
@@ -7,8 +7,8 @@ import com.ghbt.ghbt_starbucks.api.product.dto.ResponseProduct;
 import com.ghbt.ghbt_starbucks.api.product.model.Product;
 import com.ghbt.ghbt_starbucks.api.product.repository.IProductRepository;
 import com.ghbt.ghbt_starbucks.api.product_and_category.model.ProductAndCategory;
-import com.ghbt.ghbt_starbucks.api.product_search.SearchCategory;
-import com.ghbt.ghbt_starbucks.api.product_search.ISearchCategoryRepository;
+import com.ghbt.ghbt_starbucks.api.product_search.model.SearchCategory;
+import com.ghbt.ghbt_starbucks.api.product_search.repository.ISearchCategoryRepository;
 import com.ghbt.ghbt_starbucks.global.error.ServiceException;
 import com.ghbt.ghbt_starbucks.api.product.Projection.IProductSearch;
 import com.ghbt.ghbt_starbucks.api.product.Projection.IProductListByCategory;
