+++ src/main/java/com/ghbt/ghbt_starbucks/api/product_search/repository/ISearchCategoryRepository.java
@@ -1,6 +1,6 @@
-package com.ghbt.ghbt_starbucks.api.product_search;
+package com.ghbt.ghbt_starbucks.api.product_search.repository;
 
-import com.ghbt.ghbt_starbucks.api.product_search.SearchCategory;
+import com.ghbt.ghbt_starbucks.api.product_search.model.SearchCategory;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ISearchCategoryRepository extends JpaRepository<SearchCategory, Long> {
